requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 'requirements\base.in'
#
aiohttp==3.6.2
    # via discord.py
async-timeout==3.0.1
    # via aiohttp
attrs==20.3.0
    # via aiohttp
chardet==3.0.4
    # via aiohttp
discord.py==1.6.0
    # via -r requirements\base.in
idna==3.1
    # via yarl
multidict==4.7.6
    # via
    #   aiohttp
    #   yarl
typing-extensions==3.7.4.3
    # via yarl
yarl==1.6.3
    # via aiohttp
